requirements.txt
gspread>=6.0
google-auth>=2.30
google-auth-oauthlib>=1.2
google-api-python-client>=2.130
pandas>=2.2
pyarrow>=16.0
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.2
pyyaml>=6.0
rapidfuzz>=3.9
